requirements.txt
asgiref==3.8.1
blinker==1.8.2
click==8.1.7
Django==5.1.1
django-environ==0.11.2
djangorestframework==3.15.2
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysqlclient==2.2.4
sqlparse==0.5.1
Werkzeug==3.0.4
